+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [6.3.2] 2024-09-01
 
 - Open org using SF Cli + Adds internal command execution with progress
+- Fix issue when reusing an already open panel
 
 ## [6.3.1] 2024-09-01